+++ PythonVisualizations/VisualizationApp.py
@@ -73,6 +73,7 @@ class VisualizationApp(object):  # Base class for Python visualizations
     OPERATIONS_BG = 'beige'
     OPERATIONS_BORDER = 'black'
     CODE_FONT = ('Courier', 12)
+    SMALL_FONT = ('Helvetica', 9)
     CODE_HIGHLIGHT = 'yellow'
     CONTROLS_FONT = ('none', 14)
     CALL_STACK_BOUNDARY = 'brown3'
